+++ pkg/github/webhooks.go
@@ -3,12 +3,13 @@ package github
 import (
 	"net/http"
 
+	paulclient "github.com/Spazzy757/paul/pkg/client"
 	"github.com/Spazzy757/paul/pkg/helpers"
 	"github.com/google/go-github/v32/github"
 )
 
 //IncomingWebhook handles an incoming webhook request
-func IncomingWebhook(r *http.Request) error {
+func IncomingWebhook(r *http.Request, client *paulclient.GithubClient) error {
 	// handle authentication
 	secret_key := helpers.GetEnv("SECRET_KEY", "")
 	payload, validationErr := github.ValidatePayload(r, []byte(secret_key))
@@ -22,9 +23,9 @@ func IncomingWebhook(r *http.Request) error {
 	var err error
 	switch e := event.(type) {
 	case *github.IssueCommentEvent:
-		err = IssueCommentHandler(e)
+		err = IssueCommentHandler(e, client)
 	case *github.PullRequestEvent:
-		err = PullRequestHandler(e)
+		err = PullRequestHandler(e, client)
 	default:
 		break
 	}
