+++ pkg/router/router.go
@@ -4,6 +4,7 @@ import (
 	"log"
 	"net/http"
 
+	paulclient "github.com/Spazzy757/paul/pkg/client"
 	"github.com/Spazzy757/paul/pkg/github"
 	"github.com/gorilla/mux"
 )
@@ -17,7 +18,14 @@ func GetRouter() *mux.Router {
 
 //GithubWebHookHandler .
 func GithubWebHookHandler(w http.ResponseWriter, r *http.Request) {
-	err := github.IncomingWebhook(r)
+	installationIDHeader := r.Header.Get("X-GitHub-Hook-Installation-Target-ID")
+	client, err := paulclient.GetClient(installationIDHeader)
+	if err != nil {
+		log.Printf("Error: %v", err.Error())
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+	err = github.IncomingWebhook(r, client)
 	if err != nil {
 		log.Printf("Error: %v", err.Error())
 		w.WriteHeader(http.StatusBadRequest)
